PWM DC motor control W/ sonar switch
This code uses a three pin parallax "PING))) Ultrasonic Distance Sensor" in combination with a DC motor.  The PING distance is used to measure distance which is then used as a switch to turn on and off the motor at set distances. This Style switch can be used for any application requiring interaction between a stationary and mobile object triggering a given response.